Rename thread-pool.{cpp,h,hpp} to executor.{cpp,h,hpp}
