+++ c_glib/arrow-glib/meson.build
@@ -30,6 +30,7 @@ sources = files(
     'decimal.cpp',
     'decoder.cpp',
     'error.cpp',
+    'executor.cpp',
     'expression.cpp',
     'field.cpp',
     'interval.cpp',
@@ -40,7 +41,6 @@ sources = files(
     'table-builder.cpp',
     'table.cpp',
     'tensor.cpp',
-    'thread-pool.cpp',
     'timestamp-parser.cpp',
     'type.cpp',
 )
@@ -88,6 +88,7 @@ c_headers = files(
     'decimal.h',
     'decoder.h',
     'error.h',
+    'executor.h',
     'expression.h',
     'field.h',
     'interval.h',
@@ -98,7 +99,6 @@ c_headers = files(
     'table-builder.h',
     'table.h',
     'tensor.h',
-    'thread-pool.h',
     'timestamp-parser.h',
     'type.h',
 )
@@ -139,6 +139,7 @@ cpp_headers = files(
     'decimal.hpp',
     'decoder.hpp',
     'error.hpp',
+    'executor.hpp',
     'expression.hpp',
     'field.hpp',
     'interval.hpp',
@@ -149,7 +150,6 @@ cpp_headers = files(
     'table-builder.hpp',
     'table.hpp',
     'tensor.hpp',
-    'thread-pool.hpp',
     'timestamp-parser.hpp',
     'type.hpp',
 )
